Moved odc/stac/loader to odc/loader
