+++ setup.cfg
@@ -56,3 +56,5 @@ include =
 [options.package_data]
 odc.stac =
   py.typed
+odc.loader =
+  py.typed
