+++ tests/test_load.py
@@ -8,10 +8,10 @@ from dask.utils import ndeepmap
 from odc.geo.geobox import GeoBox
 from odc.geo.xr import ODCExtension
 
+from odc.loader import resolve_load_cfg
 from odc.stac import RasterLoadParams
 from odc.stac import load as stac_load
 from odc.stac._stac_load import _group_items
-from odc.stac.loader import resolve_load_cfg
 from odc.stac.testing.stac import b_, mk_parsed_item, to_stac_item
